+++ ChangeLog.md
@@ -10,6 +10,7 @@ and the format of this file is based on [Keep a Changelog](https://keepachangelo
 
 ### Added
 - Accept session_factory in `ClientBase.__init__`.
+- Helper function to iterate over pages.
 
 ### Fixed
 - Handling collections in request bodies.